# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organization Directory API",
        "description": "API for managing organizations, buildings, and activities",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Вывод информации об организации по её идентификатору",
                "description": "Returns a single organization by ID with related data",
                "operationId": "getOrganizationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of organization to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/buildings/{buildingId}/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список всех организаций находящихся в конкретном здании",
                "description": "Returns a list of organizations in a specific building",
                "operationId": "getOrganizationsByBuilding",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID of building to get organizations for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/activities/{activityId}/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Список всех организаций, которые относятся к указанному виду деятельности",
                "description": "Returns a list of organizations with a specific activity",
                "operationId": "getOrganizationsByActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID of activity to get organizations for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/organizations/search/location": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Search organizations by location",
                "description": "список организаций, которые находятся в заданном радиусе/прямоугольной области относительно указанной точки на карте. список зданий",
                "operationId": "searchOrganizationsByLocation",
                "requestBody": {
                    "description": "Search parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.7558
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.6173
                                    },
                                    "radius": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.5
                                    },
                                    "min_lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.7
                                    },
                                    "max_lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.8
                                    },
                                    "min_lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.5
                                    },
                                    "max_lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/activities/{activityId}/organizations/recursive": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "искать организации по виду деятельности. Например, поиск по виду деятельности «Еда», которая находится на первом уровне дерева, и чтобы нашлись все организации, которые относятся к видам деятельности, лежащим внутри. Т.е. в результатах поиска должны отобразиться организации с видом деятельности Еда, Мясная продукция, Молочная продукция.",
                "description": "Returns a list of organizations with a specific activity or any of its child activities",
                "operationId": "getOrganizationsByActivityRecursive",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID of activity to get organizations for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/organizations/search/name": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организации по названию",
                "description": "Returns organizations matching the provided name",
                "operationId": "searchOrganizationsByName",
                "requestBody": {
                    "description": "Search parameters",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Продукт"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "required": [
                    "name",
                    "level"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Еда"
                    },
                    "parent_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": null,
                        "nullable": true
                    },
                    "level": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Building": {
                "required": [
                    "address",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "ул. Ленина, 1"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Organization": {
                "required": [
                    "name",
                    "building_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Продуктовый рай"
                    },
                    "building_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "phone_numbers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrganizationPhone"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrganizationPhone": {
                "required": [
                    "number"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "number": {
                        "type": "string",
                        "example": "+7 (999) 123-45-67"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Organizations",
            "description": "API Endpoints for Organizations"
        }
    ]
}